Updated the joystick panel size to be almost full A3 width. Kept roughly the same height. Moved the buttons and joystick a lot further apart.
Added a template for the JLW joystick mounting plate to give me a better idea of where it will fit.

Buttons shrunk slighter to 30mm because Inkscape is a PITA with moving and centreing things. Layout is very close to a SEGA2365.

Updated the cutouts to use a rectangle with rounded corners instead of the little pieces the original AFS file used. Moved the screw holes slightly to make them match metric values. Did the same for the buttons, but this worked well since it came from a Japanese layout.